+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 - Improved tracking of hotspot event
 - Apply all events to interest grid weighted by importance
 - Apply damping to the event field of view (diagonal), as already done for location and velocity
-- Experimental/debug option to show tracking reticle
+- Options!
+  - Disable tracking (bindable)
+  - Disable user override, useful for recording
+  - Experimental/debug option to show tracking reticle
 
 ## [0.1.0] - 2023-03-22